Cleaned up implementation, added documentation + deprecations
