+++ chiselFrontend/src/main/scala/chisel3/core/MonoConnect.scala
@@ -65,8 +65,7 @@ object MonoConnect {
       connectCompileOptions: CompileOptions,
       sink: Data,
       source: Data,
-      context_mod: RawModule
-      ): Unit =
+      context_mod: RawModule): Unit =
     (sink, source) match {
 
       // Handle legal element cases, note (Bool, Bool) is caught by the first two, as Bool is a UInt
