+++ chiselFrontend/src/main/scala/chisel3/internal/Builder.scala
@@ -68,8 +68,10 @@ trait InstanceId {
   def pathName: String
   def parentPathName: String
   def parentModName: String
-  /** Returns a FIRRTL IsMember that refers to this object in the elaborated hardware graph */
+  /** Returns a FIRRTL Named that refers to this object in the elaborated hardware graph */
+  @deprecated("toNamed API is deprecated -- use toTarget instead", "3.2")
   def toNamed: Named
+  /** Returns a FIRRTL IsMember that refers to this object in the elaborated hardware graph */
   def toTarget: IsMember
 
 }
@@ -165,9 +167,21 @@ private[chisel3] trait NamedComponent extends HasId {
   /** Returns a FIRRTL ComponentName that references this object
     * @note Should not be called until circuit elaboration is complete
     */
+  @deprecated("toNamed API is deprecated -- use toTarget instead", "3.2")
   final def toNamed: ComponentName =
     ComponentName(this.instanceName, ModuleName(this.parentModName, CircuitName(this.circuitName)))
-  final def toTarget: ReferenceTarget = toNamed.toTarget
+
+  /** Returns a FIRRTL ReferenceTarget that references this object
+    * @note Should not be called until circuit elaboration is complete
+    */
+  final def toTarget: ReferenceTarget = {
+    val name = this.instanceName
+    _root_.firrtl.annotations.Target.toTargetTokens(name).toList match {
+      case Ref(r) :: components => ReferenceTarget(this.circuitName, this.parentModName, Nil, r, components)
+      case other =>
+        throw _root_.firrtl.annotations.Target.NamedException(s"Cannot convert $name into [[ReferenceTarget]]: $other")
+    }
+  }
 }
 
 // Mutable global state for chisel that can appear outside a Builder context
@@ -183,6 +197,8 @@ private[chisel3] class DynamicContext() {
   val components = ArrayBuffer[Component]()
   val annotations = ArrayBuffer[ChiselAnnotation]()
   var currentModule: Option[BaseModule] = None
+  // Contains a mapping from a elaborated module to their aspect
+  // Set by [[ModuleAspect]]
   val aspectModule: mutable.HashMap[BaseModule, BaseModule] = mutable.HashMap.empty[BaseModule, BaseModule]
   // Set by object Module.apply before calling class Module constructor
   // Used to distinguish between no Module() wrapping, multiple wrappings, and rewrapping
@@ -238,20 +254,13 @@ private[chisel3] object Builder {
   def addAspect(module: BaseModule, aspect: BaseModule): Unit = {
     dynamicContext.aspectModule += ((module, aspect))
   }
-
-
   def forcedModule: BaseModule = currentModule match {
     case Some(module) => module
-    //case None => aspectModule(this) match {
-     // case Some(module) => module
-      case None =>
-        throwException(
-          "Error: Not in a Module. Likely cause: Missed Module() wrap or bare chisel API call."
-          // A bare api call is, e.g. calling Wire() from the scala console).
-        )
-    //}
+    case None => throwException(
+      "Error: Not in a Module. Likely cause: Missed Module() wrap or bare chisel API call."
+      // A bare api call is, e.g. calling Wire() from the scala console).
+    )
   }
-
   def referenceUserModule: RawModule = {
     currentModule match {
       case Some(module: RawModule) =>
@@ -259,21 +268,18 @@ private[chisel3] object Builder {
           case Some(aspect: RawModule) => aspect
           case None => module
         }
-      case _ =>
-        throwException(
-          "Error: Not in a UserModule. Likely cause: Missed Module() wrap, bare chisel API call, or attempting to construct hardware inside a BlackBox." // scalastyle:ignore line.size.limit
-          // A bare api call is, e.g. calling Wire() from the scala console).
-        )
-    }
-  }
-
-  def forcedUserModule: RawModule = currentModule match {
-    case Some(module: RawModule) => module
-    case _ =>
-      throwException(
+      case _ => throwException(
         "Error: Not in a UserModule. Likely cause: Missed Module() wrap, bare chisel API call, or attempting to construct hardware inside a BlackBox." // scalastyle:ignore line.size.limit
         // A bare api call is, e.g. calling Wire() from the scala console).
       )
+    }
+  }
+  def forcedUserModule: RawModule = currentModule match {
+    case Some(module: RawModule) => module
+    case _ => throwException(
+      "Error: Not in a UserModule. Likely cause: Missed Module() wrap, bare chisel API call, or attempting to construct hardware inside a BlackBox." // scalastyle:ignore line.size.limit
+      // A bare api call is, e.g. calling Wire() from the scala console).
+    )
   }
   def readyForModuleConstr: Boolean = dynamicContext.readyForModuleConstr
   def readyForModuleConstr_=(target: Boolean): Unit = {
