+++ chiselFrontend/src/main/scala/chisel3/internal/firrtl/IR.scala
@@ -5,7 +5,7 @@ package chisel3.internal.firrtl
 import chisel3._
 import core._
 import chisel3.internal._
-import chisel3.internal.sourceinfo.{NoSourceInfo, SourceInfo}
+import chisel3.internal.sourceinfo.SourceInfo
 
 // scalastyle:off number.of.types
 
@@ -293,4 +293,3 @@ case class DefModule(id: RawModule, name: String, ports: Seq[Port], commands: Se
 case class DefBlackBox(id: BaseBlackBox, name: String, ports: Seq[Port], topDir: SpecifiedDirection, params: Map[String, Param]) extends Component
 
 case class Circuit(name: String, components: Seq[Component], annotations: Seq[ChiselAnnotation] = Seq.empty)
-
