--- src/main/scala/chisel3/DesignAnnotation.scala
@@ -1,5 +1,0 @@
-package chisel3
-
-import chisel3.core.{ChiselAnnotation, RawModule}
-import firrtl.annotations.{Annotation, NoTargetAnnotation}
-
